python/pip_requirements.txt
build>=0.9.0
coverage>=6.5.0
pytest>=7.2.0
pytest-cov>=4.0.0
Mako>=1.3.0
PyYAML>=5.0.0
